+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## not released
 
+- Fix: Weapons were not removed for upgrades if they are stored as plural
+- Fix: #3 Missing equipment in rule overview
+- Add: Show W6 for knocked out test
+- Add: List overview
+
+## v0.6.3 (2023-05-19)
+
+- Fix: Output spells only for army with psychic
+
+## v0.6.2 (2023-05-19)
+
+- Add: Unit Cost/Points
+
 ## v0.6.1 (2023-05-14)
 
 - Fix: Download common rules for txt import on data card creation